refactor: moved provideCodeActions to helper and addressed comments
